Moved panel class in new package
