+++ src/view/panel/TopPanel.java
@@ -1,4 +1,4 @@
-package view;
+package view.panel;
 
 import view.listener.CopyListener;
